Still for #70, seems that it needs to be an unsigned int 32, not an unsigned int 64...

+++ src/lib/geogram/basic/algorithm.h
@@ -196,7 +196,7 @@ namespace GEO {
     template <typename ITERATOR>
     inline void random_shuffle(const ITERATOR& begin, const ITERATOR& end) {
 	Numeric::int32 seed = Numeric::random_int32();
-	std::mt19937 urng{Numeric::uint64(seed)};
+	std::mt19937 urng{Numeric::uint32(seed)};
 	std::shuffle(begin, end, urng);
     }
